+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## Unrealesed
+## Unreleased
 
 * feat: app typography
+* feat: server error interceptor
+* feat: app theme
 
 ## 0.0.0+2